Extract header to a component
