+++ ui/src/app.tsx
@@ -1,51 +1,27 @@
 import { QueryClient, QueryClientProvider } from '@tanstack/react-query'
 import { ReactQueryDevtools } from '@tanstack/react-query-devtools'
-import { Breadcrumbs } from 'components/breadcrumbs'
-import { NavigationBar } from 'design-system/components/navigation/navigation-bar'
+import { Header } from 'components/header/header'
 import { BatchId } from 'pages/batch-id/batch-id'
 import { Deployments } from 'pages/deployments/deployments'
 import { Occurrences } from 'pages/occurrences/occurrences'
 import { Overview } from 'pages/overview/overview'
 import { SessionDetails } from 'pages/session-details/session-details'
 import { Sessions } from 'pages/sessions/sessions'
-import { Settings } from 'pages/settings/settings'
 import { Species } from 'pages/species/species'
 import { UnderConstruction } from 'pages/under-construction/under-construction'
-import { Navigate, Route, Routes, useNavigate } from 'react-router-dom'
+import { Navigate, Route, Routes } from 'react-router-dom'
 import { BreadcrumbContextProvider } from 'utils/breadcrumbContext'
-import { useNavItems } from 'utils/useNavItems'
 import styles from './app.module.scss'
 
 const queryClient = new QueryClient()
 
 export const App = () => {
-  const navigate = useNavigate()
-  const { navItems, activeNavItemId } = useNavItems()
-
   return (
     <QueryClientProvider client={queryClient}>
       <ReactQueryDevtools initialIsOpen={false} />
       <BreadcrumbContextProvider>
         <div className={styles.wrapper}>
-          <header className={styles.header}>
-            <div className={styles.topBar}>
-              <Breadcrumbs
-                navItems={navItems}
-                activeNavItemId={activeNavItemId}
-              />
-              <Settings />
-            </div>
-            <NavigationBar
-              items={navItems}
-              activeItemId={activeNavItemId}
-              onItemClick={(id) => {
-                const item = navItems.find((i) => i.id === id)
-                if (item) {
-                  navigate(item.path)
-                }
-              }}
-            />
-          </header>
+          <Header />
           <main className={styles.content}>
             <Routes>
               <Route path="/" element={<Navigate to="/overview" />} />
