+++ ui/src/components/header/header.module.scss
@@ -1,0 +1,41 @@
+@import 'design-system/variables/variables.scss';
+@import 'design-system/variables/colors.scss';
+
+.header {
+  position: sticky;
+  top: 0;
+  left: 0;
+  height: 140px;
+  padding: 0 100px;
+  display: flex;
+  flex-direction: column;
+  align-items: flex-start;
+  justify-content: space-between;
+  background-color: $color-neutral-50;
+  z-index: 2;
+}
+
+.topBar {
+  width: 100%;
+  height: 64px;
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+  box-sizing: border-box;
+}
+
+@media only screen and (max-width: $medium-screen-breakpoint) {
+  .header {
+    padding: 0 32px;
+  }
+}
+
+@media only screen and (max-width: $small-screen-breakpoint) {
+  .header {
+    padding: 0;
+  }
+
+  .topBar {
+    padding-left: 16px;
+  }
+}
